+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.1.0] - 2025-09-06
+
+### Added
+
+- `History`: New methods `canRedo` and `canUndo` to check if redo or undo operations are possible.
+
+### Fixed
+
+- `History`: `undo` now correctly handles empty history.
+
 ## [1.0.0] - 2025-09-01
 
 ### Added
@@ -39,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `RingBuffer`: A ring buffer (or circular buffer) data structure for TypeScript, useful for buffering data streams and implementing memory efficient queues and stacks.
 
+[1.1.0]: https://github.com/zhelvis/structure-ts/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/zhelvis/structure-ts/compare/v0.1.1...v1.0.0
 [0.1.1]: https://github.com/zhelvis/structure-ts/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/zhelvis/structure-ts/releases/tag/v0.1.0